+++ src/howard/howard/issuers/__init__.py
@@ -1,3 +1,4 @@
 """Issuer shortcuts for the howard application"""
 
+from .invoice import InvoiceDocument  # noqa: F401
 from .realisation import RealisationCertificate  # noqa: F401